Finally got the wrapper to work!

+++ ripser.h
@@ -1,3 +1,3 @@
 #include <Python.h>
 #include <numpy/arrayobject.h>
-PyArrayObject* pythondm(double* D, int N, int modulus, int dim_max, double threshold);
+PyArrayObject* pythondm(float* D, int N, int modulus, int dim_max, float threshold);
